Redesigned "Get", "Apply", "Save to file", "Load config from a local file" buttons Moved buttons to upload configs/coefficients to "Files" tab.
